+++ common/systems/src/buff.rs
@@ -370,7 +370,7 @@ fn execute_effect(
             {
                 let amount = match *kind {
                     ModifierKind::Additive => *accumulated,
-                    ModifierKind::Fractional => energy.maximum() as f32 * *accumulated,
+                    ModifierKind::Fractional => energy.maximum() * *accumulated,
                 };
                 server_emitter.emit(ServerEvent::EnergyChange {
                     entity,
@@ -395,6 +395,7 @@ fn execute_effect(
                 stat.max_energy_modifiers.mult_mod *= *value;
             },
         },
+        #[allow(clippy::manual_clamp)]
         BuffEffect::DamageReduction(dr) => {
             stat.damage_reduction = stat.damage_reduction.max(*dr).min(1.0);
         },
@@ -421,7 +422,7 @@ fn execute_effect(
                     ModifierKind::Additive => {
                         // `rate * dt` is amount of health, dividing by base max
                         // creates fraction
-                        *rate * dt / health.base_max() as f32
+                        *rate * dt / health.base_max()
                     },
                     ModifierKind::Fractional => {
                         // `rate * dt` is the fraction
@@ -466,6 +467,7 @@ fn execute_effect(
         BuffEffect::GroundFriction(gf) => {
             stat.friction_modifier *= *gf;
         },
+        #[allow(clippy::manual_clamp)]
         BuffEffect::PoiseReduction(pr) => {
             stat.poise_reduction = stat.poise_reduction.max(*pr).min(1.0);
         },
