+++ voxygen/src/window.rs
@@ -567,8 +567,7 @@ impl Window {
             // TODO: update users of this event with the fact that it is now the physical
             // size
             let winit::dpi::PhysicalSize { width, height } = physical;
-            self.events
-                .push(Event::Resize(Vec2::new(width as u32, height as u32)));
+            self.events.push(Event::Resize(Vec2::new(width, height)));
 
             // Emit event for the UI
             let logical_size = Vec2::from(Into::<(f64, f64)>::into(
